+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ This project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.htm
 ------------------
 
 - Fix `growth_factor` bug in `bracket_root()`.
+
+- Improve documentation of `bracket_root()`.
+
 1.0.1 (2020-06-15)
 ------------------