+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ Changelog
 
 This project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-1.0.2 (Unreleased)
+1.0.2 (2020-06-16)
 ------------------
 
 - Fix `growth_factor` bug in `bracket_root()`.